+++ syconn/cnn/TrainData.py
@@ -152,7 +152,7 @@ if elektronn3_avail:
             item = np.random.randint(0, len(self.sso_ids))
             self._curr_ssv_id = self.sso_ids[item]
             pts, feats = self.load_ssv_sample(item)
-            lbs = np.array([self.label_dc[self._curr_ssv_id]]*self._batch_size, dtype=np.int)
+            lbs = np.array([self.label_dc[self._curr_ssv_id]]*self._batch_size, dtype=np.int32)
             pts = torch.from_numpy(pts).float()
             lbs = torch.from_numpy(lbs[..., None]).long()
             feats = torch.from_numpy(feats).float()
@@ -282,7 +282,7 @@ if elektronn3_avail:
             assert self.train, "Other mode than 'train' is not implemented."
             self.csv_p = "/wholebrain/songbird/j0251/groundtruth/celltypes/j0251_celltype_gt_v4.csv"
             df = pandas.io.parsers.read_csv(self.csv_p, header=None, names=['ID', 'type']).values
-            ssv_ids = df[:, 0].astype(np.uint)
+            ssv_ids = df[:, 0].astype(np.uint64)
             if len(np.unique(ssv_ids)) != len(ssv_ids):
                 ixs, cnt = np.unique(ssv_ids, return_counts=True)
                 raise ValueError(f'Multi-usage of IDs! {ixs[cnt > 1]}')
@@ -338,11 +338,11 @@ if elektronn3_avail:
                 10919937, 16096256, 23144450, 2734465, 34811392, 491527,
                 15933443, 16113665, 24414208, 2854913, 37558272, 8339462,
                 15982592, 18571264, 26501121, 33581058, 46319619
-            ], dtype=np.uint)
+            ], dtype=np.uint64)
             # use celltype GT
             csv_p = '/wholebrain/songbird/j0126/GT/celltype_gt/j0126_cell_type_gt_areax_fs6_v3.csv'
             df = pandas.io.parsers.read_csv(csv_p, header=None, names=['ID', 'type']).values
-            nonglia_ssv_ids = np.concatenate([df[:, 0].astype(np.uint), nonglia_ssv_ids])
+            nonglia_ssv_ids = np.concatenate([df[:, 0].astype(np.uint64), nonglia_ssv_ids])
             self.ctx_size = ctx_size
             self.sso_params = {sso.id: dict(**sso.ssv_kwargs) for sso in ssd.get_super_segmentation_object(nonglia_ssv_ids)}
             ssd_glia = SuperSegmentationDataset(working_dir=wd_path, version='gliagt')
@@ -680,8 +680,8 @@ if elektronn3_avail:
             inp, target = self.av.getbatch(1, source='train' if self.train else 'valid')
             inp = naive_view_normalization_new(inp)
             inp, _ = self.transform(inp, None)  # Do not flip target label ^.^
-            # target = np.eye(self.ctv.n_classes)[target.squeeze().astype(np.int)]  # one-hot encoding
-            return inp[0], target.squeeze().astype(np.int)  # target should just be a scalar
+            # target = np.eye(self.ctv.n_classes)[target.squeeze().astype(np.int32)]  # one-hot encoding
+            return inp[0], target.squeeze().astype(np.int32)  # target should just be a scalar
 
         def __len__(self):
             """Determines epoch size(s)"""
@@ -720,11 +720,11 @@ if elektronn3_avail:
                 inp, target, syn_signs = self.ctv.getbatch_alternative(1, source='train' if self.train else 'valid')
                 inp, _ = self.transform(inp, None)  # Do not flip target label ^.^
                 # target should just be a scalar
-                return {'inp': (inp[0], syn_signs[0].astype(np.float32)), 'target': target.squeeze().astype(np.int)}
+                return {'inp': (inp[0], syn_signs[0].astype(np.float32)), 'target': target.squeeze().astype(np.int32)}
             else:
                 inp, target = self.ctv.getbatch_alternative_noscal(1, source='train' if self.train else 'valid')
                 inp, _ = self.transform(inp, None)  # Do not flip target label ^.^
-                return {'inp': inp[0], 'target': target.squeeze().astype(np.int)}
+                return {'inp': inp[0], 'target': target.squeeze().astype(np.int32)}
 
         def __len__(self):
             """Determines epoch size(s)"""
@@ -755,8 +755,8 @@ if elektronn3_avail:
             inp, target = self.gv.getbatch(1, source='train' if self.train else 'valid')
             inp = naive_view_normalization_new(inp)
             inp, _ = self.transform(inp, None)  # Do not flip target label ^.^
-            # target = np.eye(self.ctv.n_classes)[target.squeeze().astype(np.int)]  # one-hot encoding
-            return inp[0], target.squeeze().astype(np.int)  # target should just be a scalar
+            # target = np.eye(self.ctv.n_classes)[target.squeeze().astype(np.int32)]  # one-hot encoding
+            return inp[0], target.squeeze().astype(np.int32)  # target should just be a scalar
 
         def __len__(self):
             """Determines epoch size(s)"""
@@ -1046,8 +1046,8 @@ class MultiViewData(Data):
                 log_cnn.error(msg)
                 raise ValueError(msg)
             # TODO: Use  stratified splitting for per-class balanced splitting
-            ssv_ids = np.array(list(self.label_dict.keys()), dtype=np.uint)
-            ssv_labels = np.array(list(self.label_dict.values()), dtype=np.uint)
+            ssv_ids = np.array(list(self.label_dict.keys()), dtype=np.uint64)
+            ssv_labels = np.array(list(self.label_dict.values()), dtype=np.uint64)
             avail_classes, c_count = np.unique(ssv_labels, return_counts=True)
             n_classes = len(avail_classes)
             for c, cnt in zip(avail_classes, c_count):
@@ -1562,7 +1562,7 @@ class CelltypeViewsJ0251(CelltypeViews):
         # load GT
         self.csv_p = "/wholebrain/songbird/j0251/groundtruth/celltypes/j0251_celltype_gt_v4.csv"
         df = pandas.io.parsers.read_csv(self.csv_p, header=None, names=['ID', 'type']).values
-        ssv_ids = df[:, 0].astype(np.uint)
+        ssv_ids = df[:, 0].astype(np.uint64)
         if len(np.unique(ssv_ids)) != len(ssv_ids):
             ixs, cnt = np.unique(ssv_ids, return_counts=True)
             raise ValueError(f'Multi-usage of IDs! {ixs[cnt > 1]}')
@@ -1865,7 +1865,7 @@ class TripletData_N(Data):
         sizes = np.array([ssds_sizes[rev_dc[ix]] for ix in self.s_ids], dtype=np.float32)
         # assign weights: 0 for SV below 8, 1 for 8 to 18, 2 for 19 to 28, truncated at 5 for 49 to 58
         self.s_weights = (sizes - 8.) / 10.
-        self.s_weights *= np.array(self.s_weights > 0, dtype=np.int)
+        self.s_weights *= np.array(self.s_weights > 0, dtype=np.int32)
         self.s_weights = np.ceil(self.s_weights)
         self.s_weights[self.s_weights > 6] = 6
         print("Data Summary:\nweight\t#samples")
@@ -1942,7 +1942,7 @@ class TripletData_SSV(Data):
             ssds_sizes[i] = bb_size[i]
         # assign weights: 0 for SV below 8, 1 for 8 to 18, 2 for 19 to 28, truncated at 5 for 49 to 58
         self.sso_weights = (ssds_sizes - 8.) / 10.
-        self.sso_weights *= np.array(self.sso_weights > 0, dtype=np.int)
+        self.sso_weights *= np.array(self.sso_weights > 0, dtype=np.int32)
         self.sso_weights = np.ceil(self.sso_weights)
         self.sso_weights[self.sso_weights > 6] = 6
         self.sso_ids = np.array(self.sso_ids, dtype=np.uint32)
@@ -2050,7 +2050,7 @@ class TripletData_SSV_nviews(Data):
             ssds_sizes[i] = bb_size[i]
         # assign weights: 0 for SV below 8, 1 for 8 to 18, 2 for 19 to 28, truncated at 5 for 49 to 58
         self.sso_weights = (ssds_sizes - 8.) / 10.
-        self.sso_weights *= np.array(self.sso_weights > 0, dtype=np.int)
+        self.sso_weights *= np.array(self.sso_weights > 0, dtype=np.int32)
         self.sso_weights = np.ceil(self.sso_weights)
         self.sso_weights[self.sso_weights > 6] = 6
         self.sso_ids = np.array(self.sso_ids, dtype=np.uint32)
